Minor updates and double optimization to BigDecimal

+++ src/main/java/school/faang/supportanimals/Donation.java
@@ -2,6 +2,7 @@ package school.faang.supportanimals;
 
 import lombok.Getter;
 
+import java.math.BigDecimal;
 import java.util.concurrent.atomic.AtomicInteger;
 
 @Getter
@@ -10,9 +11,9 @@ public class Donation {
     private static final AtomicInteger ID_GENERATOR = new AtomicInteger(0);
 
     private final int id;
-    private final double amount;
+    private final BigDecimal amount;
 
-    public Donation(double amount) {
+    public Donation(BigDecimal amount) {
         this.id = ID_GENERATOR.incrementAndGet();
         this.amount = amount;
     }
